이미지 삭제시 휴지통으로 이동

+++ photo/views.py
@@ -1,7 +1,9 @@
 from django.shortcuts import render, redirect
 from PIL import Image
 from PIL.ExifTags import TAGS
-from .models import PhotoModel
+
+from user.models import UserModel
+from .models import PhotoModel, Trash
 from .forms import ImageUpload
 from .od import classification
 from django.contrib.auth.decorators import login_required
@@ -36,15 +38,37 @@ def fileUpload(request):
 
 
 @login_required
-def delete(request, photo_id):
-    photo = PhotoModel.objects.get(id=photo_id)
-    photo.delete()
-    return redirect('/')
+def img_info(request, id):
+    if request.method == 'GET':
+        photo = PhotoModel.objects.get(id=id)
+        image = PhotoModel.objects.all()
+        context = {
+            'photo': photo,
+            'img': image,
+            'id' : id,
+        }
+        return render(request, 'img_info.html', context)
+
+    elif request.method == 'POST':
+        photo = PhotoModel.objects.get(id=id)
+        trash = Trash()
+        trash.user = request.user
+        trash.trash = photo.img
+        trash.save()
+
+        photo.delete()
+      
+        return redirect('/')
 
 
 @login_required
 def trash(request):
-    return render(request, 'trash.html')
+    user = request.user.is_authenticated
+    trash = Trash.objects.all()
+    if user:
+        return render(request, 'trash.html', {'trash' : trash})
+    else:
+        return redirect('/sign-in')  
 
 
 def get_photo_info() :
@@ -92,14 +116,20 @@ def get_photo_info() :
         # 동경, 서경인지를 판단, 서경일 경우 -로 변경
         if exifGPS[3] == 'W': Lon = Lon * -1
 
-        print(Lat, ",", Lon)
-        
-def img_info(request, id):
-    if request.method == 'GET':
-        photo = PhotoModel.objects.get(id=id)
-        image = PhotoModel.objects.all()
-    
-    return render(request, 'img_info.html', context=dict(photo=photo, img=image, id=id))
+        context = {
+            'DateTime': DateTime,
+            'ExifImageHeight': ExifImageHeight,
+            'ExifImageWidth': ExifImageWidth,
+            'ShutterSpeedValue': ShutterSpeedValue,
+            'FNumber': FNumber,
+            'Make': Make,
+            'Model': Model,
+            'LensModel': LensModel,
+            'Lat' : Lat,
+            'Lon' : Lon,
+        }
+
+        return render(request, img_info.html, context)
 
 
 # 즐겨찾기
